feat: allow adding annotations to the Service
Example use case: add the "prometheus.io/scrape: true" annotation on the
service (not the pods themselves).
Directly scrapping the service allows not to have new series whenever the
exporter pod is recreated with a new name, as series scrapped from pods
typically include the pod name in their labels.